+++ src/components/Popup/CallPopup/CallPopup.tsx
@@ -5,7 +5,7 @@ import Image from 'next/image'
 import bgLogo from '@/assets/static/call-popup-bg.svg'
 import Link from 'next/link'
 import ClosePopupButton from '@/ui/buttons/ClosePopupButton'
-import { callFormHandler } from '@/features/forms'
+import { callFormHandler } from '@/features/actions/call-form'
 import Form from '@/ui/Form'
 import { getSocials } from '@/globals/api/methods/getSocials'
 import { socialItem } from '@/layout/Footer/Footer'
